requirements.txt
python-telegram-bot
requests
unidecode